+++ sewup/src/kv/traits/mod.rs
@@ -4,3 +4,6 @@ pub use key::Key;
 
 pub mod value;
 pub use value::Value;
+
+pub mod vec;
+pub use vec::VecLike;
